> move storage in /common > outsource changeWattInKiloWatt
