+++ ui/src/app/edge/live/common/storage/modal/modal.component.html
